+++ MROZA-Show/MROZA-Show-ejb/src/test/java/com/mroza/service/SyncServiceDbImplTest.java
@@ -124,19 +124,19 @@ public class SyncServiceDbImplTest {
 
         KidTable updateKidTable = new KidTable(true, true, updateTable, updatePeriod);
         updateKidTable.setId(kidTableToUpdate.getId());
-        updateKidTable.setTableId(updateTable.getId());
-        updateKidTable.setPeriodId(updatePeriod.getId());
+        updateKidTable.setTableId(updateTable.getId().intValue());
+        updateKidTable.setPeriodId(updatePeriod.getId().intValue());
         updateKidTable.setNote("Nowa notatka");
         updateKidTable.setLastModDate(mockModDatetime);
         List<ResolvedField> updateResolvedFields = updateKidTable.getResolvedFields();
         updateResolvedFields.get(0).setValue("OK");
         updateResolvedFields.get(0).setId(kidTableToUpdate.getResolvedFields().get(0).getId());
-        updateResolvedFields.get(0).setTableFieldId(kidTableToUpdate.getResolvedFields().get(0).getTableField().getId());
-        updateResolvedFields.get(0).setKidTableId(kidTableToUpdate.getId());
+        updateResolvedFields.get(0).setTableFieldId(kidTableToUpdate.getResolvedFields().get(0).getTableField().getId().intValue());
+        updateResolvedFields.get(0).setKidTableId(kidTableToUpdate.getId().intValue());
         updateResolvedFields.get(1).setValue("NOK");
         updateResolvedFields.get(1).setId(kidTableToUpdate.getResolvedFields().get(1).getId());
-        updateResolvedFields.get(1).setTableFieldId(kidTableToUpdate.getResolvedFields().get(1).getTableField().getId());
-        updateResolvedFields.get(1).setKidTableId(kidTableToUpdate.getId());
+        updateResolvedFields.get(1).setTableFieldId(kidTableToUpdate.getResolvedFields().get(1).getTableField().getId().intValue());
+        updateResolvedFields.get(1).setKidTableId(kidTableToUpdate.getId().intValue());
 
         ReceiveSyncModel receiveSyncModel = new ReceiveSyncModel();
         receiveSyncModel.getTransferChildTableList().add(TransferChildTable.transferObjectFromServerModel(updateKidTable));
@@ -146,14 +146,14 @@ public class SyncServiceDbImplTest {
         syncService.updateModelWithDataFromAndroid(receiveSyncModel);
         sqlSession.commit();
 
-        Assert.assertEquals("Nowa notatka", kidTablesDao.selectKidTableById(kidTableToUpdate.getId()).getNote());
-        Assert.assertEquals(true, kidTablesDao.selectKidTableById(kidTableToUpdate.getId()).isCollectingLearning());
-        Assert.assertEquals(true, kidTablesDao.selectKidTableById(kidTableToUpdate.getId()).isCollectingGeneralization());
-        Assert.assertEquals(mockModDatetime, kidTablesDao.selectKidTableById(kidTableToUpdate.getId()).getLastModDate());
+        Assert.assertEquals("Nowa notatka", kidTablesDao.selectKidTableById(kidTableToUpdate.getId().intValue()).getNote());
+        Assert.assertEquals(true, kidTablesDao.selectKidTableById(kidTableToUpdate.getId().intValue()).isCollectingLearning());
+        Assert.assertEquals(true, kidTablesDao.selectKidTableById(kidTableToUpdate.getId().intValue()).isCollectingGeneralization());
+        Assert.assertEquals(mockModDatetime, kidTablesDao.selectKidTableById(kidTableToUpdate.getId().intValue()).getLastModDate());
         Assert.assertEquals(2, kidTablesDao.selectAllKidTables().size());
 
-        Assert.assertEquals("OK", resolvedFieldsDao.selectResolvedFieldById(updateResolvedFields.get(0).getId()).getValue());
-        Assert.assertEquals(kidTableToUpdate.getId(), resolvedFieldsDao.selectResolvedFieldById(updateResolvedFields.get(0).getId()).getKidTableId());
+        Assert.assertEquals("OK", resolvedFieldsDao.selectResolvedFieldById(updateResolvedFields.get(0).getId().intValue()).getValue());
+        Assert.assertEquals(kidTableToUpdate.getId(), resolvedFieldsDao.selectResolvedFieldById(updateResolvedFields.get(0).getId().intValue()).getKidTableId());
 
     }
